Extract getRootIdentifier to shared AST utilities
