+++ packages/utilities/ast/src/index.ts
@@ -11,8 +11,9 @@ export * from "./function-init-path";
 export * from "./function-is";
 export * from "./identifier-is";
 export * from "./identifier-name";
+export * from "./identifier-traverse";
 export * from "./literal-is";
-export * from "./node-equal";
+export * from "./node-equality";
 export * from "./node-is";
 export * from "./node-kind";
 export * from "./node-line";
